Refactor: 파일명 및 변수명 수정
- 이미지 변수명 카멜케이스로 변환
- 파일명 직관적으로 이해할 수 있게 변환
