+++ src/components/ProfileImg/ProfileImgOption.jsx
@@ -1,6 +1,6 @@
-import * as S from './ProfileImgList.style';
+import * as S from './ProfileImgOption.style';
 
-export default function ProfileImgList() {
+export default function ProfileImgOption() {
   return (
     <S.ProfileListContainer>
       <S.Subtitle>프로필 이미지를 선택해주세요!</S.Subtitle>
